# [discussion] Hacky experiment to optimize sum()