Disclaimer: this is not a PR to propose a change, just to have a conversation on it.

In https://github.com/prometheus/prometheus/pull/8594 I've proposed some changes to speed up the aggregation function. While working on it I was wondering if there may be a more optimized way to run the `rangeEval()` with an aggregation.

I'm not a PromQL expert so I may miss something very obvious. As a pure experiment, I've implemented a simplified `rangeEvalSum()` which is the `rangeEval()` + `SUM` aggregation and below you can find the result of the range query benchmark.

My intention is not to propose `rangeEvalSum()` to Prometheus but starting from a specific case (the very common sum aggregation) to investigate if we can optimize it and then try to do the same with other aggregations. Before doing that, I would like to assess correctness.

Question: is `rangeEvalSum()` correct or not?

## Range query benchmarks for queries using `sum()`

```
name                                                                                                       old time/op    new time/op    delta
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                 8.39ms ± 1%    5.90ms ± 1%  -29.60%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                      172ms ±10%      68ms ± 2%  -60.18%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                     186ms ± 1%      69ms ± 0%  -63.07%  (p=0.016 n=5+4)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                           182ms ± 2%      69ms ± 2%  -62.11%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                          172ms ±11%      68ms ± 0%  -60.19%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                           18.0ms ± 0%    13.8ms ± 1%  -23.75%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12    36.4ms ± 0%    28.1ms ± 0%  -22.96%  (p=0.008 n=5+5)

name                                                                                                       old alloc/op   new alloc/op   delta
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                  795kB ± 0%     345kB ± 0%  -56.58%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                     7.31MB ± 0%    3.83MB ± 0%  -47.68%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                    33.7MB ± 0%     5.3MB ± 0%  -84.26%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                          27.3MB ± 0%     5.3MB ± 0%  -80.60%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                         6.61MB ± 0%    3.83MB ± 0%  -42.08%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                            868kB ± 0%     360kB ± 0%  -58.52%  (p=0.016 n=4+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12    1.90MB ± 0%    0.88MB ± 0%  -53.55%  (p=0.008 n=5+5)

name                                                                                                       old allocs/op  new allocs/op  delta
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                  12.9k ± 0%      5.3k ± 0%  -58.62%  (p=0.000 n=5+4)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                       115k ± 0%       57k ± 0%  -50.01%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                      396k ± 0%       58k ± 0%  -85.40%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                            396k ± 0%       58k ± 0%  -85.40%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                           115k ± 0%       57k ± 0%  -50.01%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                            14.5k ± 0%      5.4k ± 0%  -63.03%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12     33.0k ± 0%     14.7k ± 0%  -55.45%  (p=0.016 n=5+4)
```