@roidelapluie 2021-03-15
Please look at https://github.com/prometheus/prometheus/pull/3966 which was the last big promql refactoring.