@pracucci 2021-03-15
_Briefly discussed with Julien offline and mentioned that this experiment may be very bad with churning series (and I agree). I will add some benchmarks to try to measure it._