@pracucci 2021-03-17
I've tested this PR sending few selected `sum()` queries to a dedicated pool of Cortex queries running this PR and, from very preliminary results, the performance benefit looks significative (but such queries have nearly zero churning series). Latency has been cut by about -60%, which matches with the benchmark posted above.