@pracucci 2021-03-24
I've posted updated benchmarks in the PR description.