@roidelapluie 2021-03-24
It seems that you are handling NaN's in a special way by scanning the code. NaN can be rendered by PromQL. In fact the following code panics:

```
# Tests for NaN.
clear
load 5m
  http_requests     NaN

eval instant at 0m sum(http_requests)
  {} NaN
```

```
--- FAIL: TestEvaluations (0.05s)                                    
    --- FAIL: TestEvaluations/testdata/aggregators.test (0.05s)
panic: runtime error: index out of range [0] with length 0 [recovered]
        panic: runtime error: index out of range [0] with length 0
                                                                     
goroutine 48 [running]:
testing.tRunner.func1.2(0x9715c0, 0xc0026c4090)
        /home/roidelapluie/godist/go/src/testing/testing.go:1144 +0x332
testing.tRunner.func1(0xc001f0a300)
        /home/roidelapluie/godist/go/src/testing/testing.go:1147 +0x4b6
panic(0x9715c0, 0xc0026c4090)
        /home/roidelapluie/godist/go/src/runtime/panic.go:965 +0x1b9
github.com/prometheus/prometheus/promql.(*Engine).execEvalStmt(0xc001135340, 0xa4d040, 0xc0002fe3f0, 0xc00113daa0, 0xc001144140, 0x0, 0x0, 
0x0, 0x0, 0x0, ...)                                                  
        /home/roidelapluie/dev/prometheus/promql/engine.go:610 +0x148f
github.com/prometheus/prometheus/promql.(*Engine).exec(0xc001135340, 0xa4d040, 0xc0002fe3f0, 0xc00113daa0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, ..
.)                                
        /home/roidelapluie/dev/prometheus/promql/engine.go:536 +0x4fe 
github.com/prometheus/prometheus/promql.(*query).Exec(0xc00113daa0, 0xa4cf98, 0xc0011485c0, 0xc001f14bc4)
        /home/roidelapluie/dev/prometheus/promql/engine.go:174 +0x9d
github.com/prometheus/prometheus/promql.(*Test).exec(0xc00047e320, 0xa44720, 0xc00047f310, 0x0, 0x0)
        /home/roidelapluie/dev/prometheus/promql/test.go:537 +0x2e2
github.com/prometheus/prometheus/promql.(*Test).Run(0xc00047e320, 0xc001f0a300, 0x0)
        /home/roidelapluie/dev/prometheus/promql/test.go:433 +0x71
github.com/prometheus/prometheus/promql.TestEvaluations.func1(0xc001f0a300)
        /home/roidelapluie/dev/prometheus/promql/promql_test.go:31 +0xa7
testing.tRunner(0xc001f0a300, 0xc001e82af0)
        /home/roidelapluie/godist/go/src/testing/testing.go:1194 +0xef
created by testing.(*T).Run
        /home/roidelapluie/godist/go/src/testing/testing.go:1239 +0x2b3
FAIL    github.com/prometheus/prometheus/promql 0.230s
FAIL
```