@pracucci 2021-03-25
> It seems that you are handling NaN's in a special way by scanning the code. NaN can be rendered by PromQL. In fact the following code panics:

Thanks @roidelapluie for pointing out. Yes, I was doing an hack with NaN. I should have fixed it in the last commit.