@pracucci 2022-08-12
**Updated benchmarks**:

```
name                                                                                                       old time/op    new time/op    delta
RangeQuery/expr=sum(a_one),steps=1-12                                                                        17.0µs ± 0%    17.5µs ± 1%   +3.11%
RangeQuery/expr=sum(a_one),steps=10-12                                                                       21.3µs ± 1%    16.4µs ± 2%  -23.22%
RangeQuery/expr=sum(a_one),steps=100-12                                                                      66.8µs ± 1%    22.1µs ± 1%  -66.83%
RangeQuery/expr=sum(a_one),steps=1000-12                                                                      500µs ± 1%      69µs ± 2%  -86.26%
RangeQuery/expr=sum(a_ten),steps=1-12                                                                        50.1µs ± 2%    47.0µs ± 1%   -6.14%
RangeQuery/expr=sum(a_ten),steps=10-12                                                                       58.4µs ± 1%    51.4µs ± 1%  -12.04%
RangeQuery/expr=sum(a_ten),steps=100-12                                                                       170µs ± 1%     105µs ± 1%  -38.21%
RangeQuery/expr=sum(a_ten),steps=1000-12                                                                     1.19ms ± 1%    0.50ms ± 2%  -57.95%
RangeQuery/expr=sum(a_hundred),steps=1-12                                                                     371µs ± 1%     360µs ± 1%   -2.85%
RangeQuery/expr=sum(a_hundred),steps=10-12                                                                    423µs ± 1%     391µs ± 1%   -7.54%
RangeQuery/expr=sum(a_hundred),steps=100-12                                                                  1.13ms ± 2%    0.92ms ± 2%  -19.15%
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                 7.61ms ± 7%    4.75ms ± 2%  -37.60%
RangeQuery/expr=sum_without_(l)(h_one),steps=1-12                                                            64.9µs ± 5%    53.9µs ± 1%  -17.02%
RangeQuery/expr=sum_without_(l)(h_one),steps=10-12                                                            107µs ± 1%      58µs ± 1%  -45.65%
RangeQuery/expr=sum_without_(l)(h_one),steps=100-12                                                           560µs ± 0%     123µs ± 1%  -78.08%
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-12                                                         4.99ms ± 1%    0.62ms ± 3%  -87.62%
RangeQuery/expr=sum_without_(l)(h_ten),steps=1-12                                                             404µs ± 1%     383µs ± 1%   -5.32%
RangeQuery/expr=sum_without_(l)(h_ten),steps=10-12                                                            505µs ± 4%     418µs ± 1%  -17.15%
RangeQuery/expr=sum_without_(l)(h_ten),steps=100-12                                                          1.70ms ± 2%    0.97ms ± 1%  -42.85%
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-12                                                         12.8ms ± 1%     5.3ms ± 3%  -58.69%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1-12                                                        4.10ms ± 2%    3.85ms ± 6%   -6.00%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=10-12                                                       4.80ms ± 2%    4.24ms ± 1%  -11.60%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-12                                                      15.1ms ± 2%    10.7ms ± 3%  -29.52%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                      105ms ± 1%      53ms ± 1%  -49.30%
RangeQuery/expr=sum_without_(le)(h_one),steps=1-12                                                           53.0µs ± 1%    50.6µs ± 1%   -4.44%
RangeQuery/expr=sum_without_(le)(h_one),steps=10-12                                                          63.0µs ± 2%    53.6µs ± 0%  -14.98%
RangeQuery/expr=sum_without_(le)(h_one),steps=100-12                                                          178µs ± 1%     111µs ± 0%  -37.84%
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-12                                                        1.23ms ± 0%    0.54ms ± 1%  -56.21%
RangeQuery/expr=sum_without_(le)(h_ten),steps=1-12                                                            397µs ± 0%     381µs ± 0%   -3.97%
RangeQuery/expr=sum_without_(le)(h_ten),steps=10-12                                                           497µs ± 2%     412µs ± 1%  -16.99%
RangeQuery/expr=sum_without_(le)(h_ten),steps=100-12                                                         1.65ms ± 1%    0.99ms ± 1%  -39.83%
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-12                                                        12.0ms ± 1%     5.3ms ± 6%  -55.71%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1-12                                                       4.23ms ± 2%    3.85ms ± 1%   -9.09%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=10-12                                                      5.27ms ± 4%    4.22ms ± 2%  -20.06%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-12                                                     20.2ms ± 3%    11.0ms ± 3%  -45.80%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                     131ms ± 2%      54ms ± 1%  -59.07%
RangeQuery/expr=sum_by_(l)(h_one),steps=1-12                                                                 52.1µs ± 1%    50.6µs ± 2%   -2.95%
RangeQuery/expr=sum_by_(l)(h_one),steps=10-12                                                                61.9µs ± 1%    54.0µs ± 0%  -12.76%
RangeQuery/expr=sum_by_(l)(h_one),steps=100-12                                                                181µs ± 2%     111µs ± 1%  -38.71%
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-12                                                              1.25ms ± 3%    0.54ms ± 1%  -57.14%
RangeQuery/expr=sum_by_(l)(h_ten),steps=1-12                                                                  397µs ± 1%     382µs ± 3%   -3.76%
RangeQuery/expr=sum_by_(l)(h_ten),steps=10-12                                                                 488µs ± 1%     414µs ± 1%  -15.19%
RangeQuery/expr=sum_by_(l)(h_ten),steps=100-12                                                               1.64ms ± 2%    0.98ms ± 2%  -40.10%
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-12                                                              12.1ms ± 0%     5.2ms ± 2%  -57.11%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1-12                                                             4.22ms ± 4%    3.82ms ± 1%   -9.50%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=10-12                                                            5.30ms ± 5%    4.24ms ± 2%  -20.13%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-12                                                           20.3ms ± 2%    10.9ms ± 3%  -46.11%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                           133ms ± 0%      54ms ± 2%  -59.15%
RangeQuery/expr=sum_by_(le)(h_one),steps=1-12                                                                63.0µs ± 1%    54.0µs ± 1%  -14.35%
RangeQuery/expr=sum_by_(le)(h_one),steps=10-12                                                                107µs ± 2%      58µs ± 2%  -45.49%
RangeQuery/expr=sum_by_(le)(h_one),steps=100-12                                                               554µs ± 1%     123µs ± 1%  -77.85%
RangeQuery/expr=sum_by_(le)(h_one),steps=1000-12                                                             4.99ms ± 3%    0.60ms ± 1%  -87.94%
RangeQuery/expr=sum_by_(le)(h_ten),steps=1-12                                                                 402µs ± 1%     380µs ± 0%   -5.46%
RangeQuery/expr=sum_by_(le)(h_ten),steps=10-12                                                                501µs ± 1%     420µs ± 1%  -16.08%
RangeQuery/expr=sum_by_(le)(h_ten),steps=100-12                                                              1.73ms ± 1%    0.99ms ± 2%  -43.17%
RangeQuery/expr=sum_by_(le)(h_ten),steps=1000-12                                                             12.7ms ± 3%     5.4ms ± 5%  -57.94%
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1-12                                                            4.11ms ± 1%    3.99ms ± 6%   -3.07%
RangeQuery/expr=sum_by_(le)(h_hundred),steps=10-12                                                           4.69ms ± 1%    4.23ms ± 2%   -9.88%
RangeQuery/expr=sum_by_(le)(h_hundred),steps=100-12                                                          14.9ms ± 5%    10.9ms ± 3%  -26.85%
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-12                                                          105ms ± 2%      53ms ± 2%  -49.92%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1-12                                                  24.1µs ± 1%    23.9µs ± 4%   -0.91%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=10-12                                                 29.2µs ± 1%    24.2µs ± 2%  -17.34%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=100-12                                                79.1µs ± 2%    37.0µs ± 1%  -53.21%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m])),steps=1000-12                                                553µs ± 1%     148µs ± 1%  -73.29%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1-12                                                  59.2µs ± 1%    55.7µs ± 2%   -5.87%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=10-12                                                 74.0µs ± 0%    66.1µs ± 2%  -10.64%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=100-12                                                 253µs ± 1%     189µs ± 1%  -25.45%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m])),steps=1000-12                                               1.90ms ± 1%    1.24ms ± 4%  -34.91%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1-12                                               399µs ± 1%     388µs ± 2%   -2.82%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=10-12                                              505µs ± 1%     475µs ± 1%   -6.02%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-12                                            1.91ms ± 2%    1.68ms ± 1%  -12.01%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-12                                           14.3ms ± 2%    11.7ms ± 2%  -18.40%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1-12               44.7µs ± 0%    42.0µs ± 2%   -6.02%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=10-12              57.5µs ± 1%    47.1µs ± 1%  -18.08%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=100-12              187µs ± 2%     103µs ± 3%  -45.14%
RangeQuery/expr=sum_without_(l)(rate(a_one[1m]))_/_sum_without_(l)(rate(b_one[1m])),steps=1000-12            1.42ms ± 3%    0.60ms ± 1%  -57.98%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1-12                115µs ± 1%     110µs ± 1%   -4.13%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=10-12               148µs ± 1%     133µs ± 0%  -10.52%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=100-12              533µs ± 0%     409µs ± 2%  -23.23%
RangeQuery/expr=sum_without_(l)(rate(a_ten[1m]))_/_sum_without_(l)(rate(b_ten[1m])),steps=1000-12            4.04ms ± 0%    2.69ms ± 1%  -33.48%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1-12        788µs ± 2%     774µs ± 2%   -1.73%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=10-12      1.01ms ± 1%    0.98ms ± 5%   -3.27%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=100-12     3.88ms ± 3%    3.41ms ± 2%  -12.04%
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-12    29.0ms ± 2%    24.0ms ± 1%  -17.32%
RangeQuery_HighCardinalityAggregation-12                                                                      5.91s ± 2%     0.51s ± 3%  -91.45%

name                                                                                                       old alloc/op   new alloc/op   delta
RangeQuery/expr=sum(a_one),steps=1-12                                                                        8.21kB ± 0%    7.17kB ± 0%  -12.67%
RangeQuery/expr=sum(a_one),steps=10-12                                                                       12.0kB ± 0%     7.5kB ± 0%  -37.32%
RangeQuery/expr=sum(a_one),steps=100-12                                                                      50.5kB ± 0%    11.7kB ± 0%  -76.76%
RangeQuery/expr=sum(a_one),steps=1000-12                                                                      439kB ± 0%      53kB ± 0%  -87.99%
RangeQuery/expr=sum(a_ten),steps=1-12                                                                        19.7kB ± 0%    17.1kB ± 0%  -12.97%
RangeQuery/expr=sum(a_ten),steps=10-12                                                                       23.8kB ± 0%    17.5kB ± 0%  -26.43%
RangeQuery/expr=sum(a_ten),steps=100-12                                                                      67.0kB ± 0%    23.6kB ± 0%  -64.82%
RangeQuery/expr=sum(a_ten),steps=1000-12                                                                      533kB ± 0%      93kB ± 0%  -82.47%
RangeQuery/expr=sum(a_hundred),steps=1-12                                                                     130kB ± 0%     112kB ± 0%  -13.59%
RangeQuery/expr=sum(a_hundred),steps=10-12                                                                    134kB ± 0%     113kB ± 0%  -15.96%
RangeQuery/expr=sum(a_hundred),steps=100-12                                                                   196kB ± 0%     138kB ± 0%  -29.86%
RangeQuery/expr=sum(a_hundred),steps=1000-12                                                                 1.14MB ± 0%    0.48MB ± 0%  -57.73%
RangeQuery/expr=sum_without_(l)(h_one),steps=1-12                                                            26.5kB ± 0%    20.4kB ± 0%  -23.00%
RangeQuery/expr=sum_without_(l)(h_one),steps=10-12                                                           51.3kB ± 0%    23.6kB ± 0%  -53.97%
RangeQuery/expr=sum_without_(l)(h_one),steps=100-12                                                           302kB ± 0%      61kB ± 0%  -79.84%
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-12                                                         2.85MB ± 0%    0.43MB ± 0%  -85.01%
RangeQuery/expr=sum_without_(l)(h_ten),steps=1-12                                                             159kB ± 0%     136kB ± 0%  -14.35%
RangeQuery/expr=sum_without_(l)(h_ten),steps=10-12                                                            187kB ± 0%     140kB ± 0%  -25.44%
RangeQuery/expr=sum_without_(l)(h_ten),steps=100-12                                                           491kB ± 0%     198kB ± 0%  -59.73%
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-12                                                         3.89MB ± 0%    0.88MB ± 0%  -77.35%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1-12                                                        1.46MB ± 0%    1.27MB ± 0%  -13.19%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=10-12                                                       1.49MB ± 0%    1.27MB ± 0%  -14.60%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-12                                                      2.00MB ± 0%    1.53MB ± 0%  -23.26%
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-12                                                     10.7MB ± 0%     5.3MB ± 0%  -50.92%
RangeQuery/expr=sum_without_(le)(h_one),steps=1-12                                                           21.1kB ± 0%    18.3kB ± 0%  -13.11%
RangeQuery/expr=sum_without_(le)(h_one),steps=10-12                                                          25.2kB ± 0%    18.7kB ± 0%  -25.82%
RangeQuery/expr=sum_without_(le)(h_one),steps=100-12                                                         68.6kB ± 0%    24.9kB ± 0%  -63.63%
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-12                                                         540kB ± 0%      98kB ± 0%  -81.85%
RangeQuery/expr=sum_without_(le)(h_ten),steps=1-12                                                            159kB ± 0%     136kB ± 0%  -14.23%
RangeQuery/expr=sum_without_(le)(h_ten),steps=10-12                                                           185kB ± 0%     139kB ± 0%  -24.78%
RangeQuery/expr=sum_without_(le)(h_ten),steps=100-12                                                          470kB ± 0%     194kB ± 0%  -58.73%
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-12                                                        3.68MB ± 0%    0.85MB ± 0%  -76.99%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1-12                                                       1.52MB ± 0%    1.29MB ± 0%  -15.10%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=10-12                                                      1.77MB ± 0%    1.32MB ± 0%  -25.71%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-12                                                     4.56MB ± 0%    1.86MB ± 0%  -59.27%
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-12                                                    36.0MB ± 0%     8.2MB ± 0%  -77.28%
RangeQuery/expr=sum_by_(l)(h_one),steps=1-12                                                                 21.0kB ± 0%    18.3kB ± 0%  -13.17%
RangeQuery/expr=sum_by_(l)(h_one),steps=10-12                                                                25.1kB ± 0%    18.6kB ± 0%  -25.87%
RangeQuery/expr=sum_by_(l)(h_one),steps=100-12                                                               68.6kB ± 0%    24.9kB ± 0%  -63.67%
RangeQuery/expr=sum_by_(l)(h_one),steps=1000-12                                                               540kB ± 0%      98kB ± 0%  -81.86%
RangeQuery/expr=sum_by_(l)(h_ten),steps=1-12                                                                  159kB ± 0%     136kB ± 0%  -14.50%
RangeQuery/expr=sum_by_(l)(h_ten),steps=10-12                                                                 185kB ± 0%     139kB ± 0%  -25.01%
RangeQuery/expr=sum_by_(l)(h_ten),steps=100-12                                                                470kB ± 0%     193kB ± 0%  -58.82%
RangeQuery/expr=sum_by_(l)(h_ten),steps=1000-12                                                              3.68MB ± 0%    0.85MB ± 0%  -77.00%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1-12                                                             1.52MB ± 0%    1.28MB ± 0%  -15.35%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=10-12                                                            1.77MB ± 0%    1.31MB ± 0%  -25.94%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-12                                                           4.56MB ± 0%    1.85MB ± 0%  -59.36%
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-12                                                          36.0MB ± 0%     8.2MB ± 0%  -77.29%
RangeQuery/expr=sum_by_(le)(h_one),steps=1-12                                                                26.5kB ± 0%    20.3kB ± 0%  -23.38%
…